targetTickRate as integer

+++ lib/globals.dart
@@ -1,5 +1,5 @@
 class Globals {
   static double screenWidth = 400;
   static double screenHeight = 800;
-  static double targetTickRate = 30
+  static int targetTickRate = 30
 }
